+++ .travis.yml
@@ -1,6 +1,5 @@
 language: python
 python:
-    - "2.7"
     - "3.4"
     - "3.5"
     - "3.6"
