+++ __init__.py
@@ -1,0 +1,7 @@
+# this __init__.py is only meant for local package development
+
+try:
+    from .arpeggio import *
+    # this we need for pip install --install-option test
+except ImportError:
+    import arpeggio
