+++ arpeggio/tests/regressions/issue_26/test_issue_26.py
@@ -1,4 +1,4 @@
-from arpeggio.cleanpeg import ParserPEG
+from ....cleanpeg import ParserPEG
 
 def test_regex_with_empty_successful_match_in_repetition():
     grammar = \
