+++ arpeggio/tests/regressions/issue_32/test_issue_32.py
@@ -1,18 +1,15 @@
-# -*- coding: utf-8 -*-
-
 # Test github issue 32: ensure that Python-style escape sequences in peg and
 # cleanpeg grammars are properly converted, and ensure that escaping of those
 # sequences works as well.
 
-from __future__ import print_function
 import re
 import sys
 
 import pytest
 
-import arpeggio
-from arpeggio.cleanpeg import ParserPEG as ParserCleanPEG
-from arpeggio.peg import ParserPEG
+from .... import *
+from ....cleanpeg import ParserPEG as ParserCleanPEG
+from ....peg import ParserPEG
 
 
 def check_parser(grammar, text):
