+++ arpeggio/tests/regressions/test_memoization.py
@@ -1,7 +1,6 @@
-from __future__ import unicode_literals
 import pytest
 import sys
-from arpeggio import ParserPython
+from .... import ParserPython
 
 
 def test_memoization_positive(capsys):
