+++ arpeggio/tests/test_decorator_combine.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 #######################################################################
 # Name: test_decorator_combine
 # Purpose: Test for Combine decorator. Combine decorator
@@ -9,9 +8,8 @@
 # License: MIT License
 #######################################################################
 
-from __future__ import unicode_literals
 import pytest
-from arpeggio import ParserPython, ZeroOrMore, OneOrMore, NonTerminal, \
+from .. import ParserPython, ZeroOrMore, OneOrMore, NonTerminal, \
     Terminal, NoMatch, Combine
