+++ arpeggio/tests/test_error_reporting.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 #######################################################################
 # Name: test_error_reporting
 # Purpose: Test error reporting for various cases.
@@ -6,11 +5,11 @@
 # Copyright: (c) 2015 Igor R. Dejanović <igor DOT dejanovic AT gmail DOT com>
 # License: MIT License
 #######################################################################
-from __future__ import unicode_literals
+
 import pytest
 
-from arpeggio import Optional, Not, ParserPython, NoMatch, EOF
-from arpeggio import RegExMatch as _
+from .. import Optional, Not, ParserPython, NoMatch, EOF
+from .. import RegExMatch as _
 
 
 def test_non_optional_precedence():
