+++ arpeggio/tests/test_parser_resilience.py
@@ -1,6 +1,4 @@
-# coding=utf-8
-from __future__ import unicode_literals, absolute_import
-from arpeggio import ParserPython, EOF
+from .. import ParserPython, EOF
 import pytest
