+++ arpeggio/tests/test_suppression.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 #######################################################################
 # Name: test_suppression
 # Purpose: Test suppresion of parse tree nodes.
@@ -7,9 +6,8 @@
 # License: MIT License
 #######################################################################
 
-from __future__ import unicode_literals
 import pytest  # noqa
-from arpeggio import ParserPython, Sequence
+from .. import ParserPython, Sequence
 
 
 def test_sequence_suppress():
